+++ packages/suite/src/actions/suite/__fixtures__/suiteActions.ts
@@ -1,5 +1,6 @@
 import { DEVICE, TRANSPORT } from '@trezor/connect';
 import { SUITE, NOTIFICATION, MODAL } from '@suite-actions/constants';
+import { init as connectInit } from '@suite-actions/trezorConnectActions';
 import { DISCOVERY } from '@wallet-actions/constants';
 import { TorStatus } from '@suite-types';
 import * as suiteActions from '../suiteActions';
@@ -730,6 +731,16 @@ const observeSelectedDevice = [
 ];
 
 const acquireDevice = [
+    {
+        description: `prepare init pending`,
+        state: {},
+        result: connectInit.pending.type,
+    },
+    {
+        description: `prepare init fulfilled`,
+        state: {},
+        result: connectInit.fulfilled.type,
+    },
     {
         description: `success`,
         state: {
