+++ packages/suite/src/actions/suite/__tests__/suiteActions.test.ts
@@ -15,7 +15,6 @@ import { SUITE } from '../constants';
 import * as suiteActions from '../suiteActions';
 import { init } from '../trezorConnectActions';
 import fixtures from '../__fixtures__/suiteActions';
-import { discardMockedConnectInitActions } from '@suite-utils/storage';
 
 const { getSuiteDevice } = global.JestMocks;
 
@@ -218,11 +217,10 @@ describe('Suite Actions', () => {
             const store = initStore(state);
             store.dispatch(init()); // trezorConnectActions.init needs to be called in order to wrap "getFeatures" with lockUi action
             await store.dispatch(suiteActions.acquireDevice(f.requestedDevice));
-            const expectedActions = discardMockedConnectInitActions(store.getActions());
             if (!f.result) {
-                expect(expectedActions.length).toEqual(0);
+                expect(store.getActions().length).toEqual(0);
             } else {
-                const action = expectedActions.pop();
+                const action = store.getActions().pop();
                 expect(action.type).toEqual(f.result);
             }
         });
