+++ packages/suite/src/utils/suite/storage.ts
@@ -1,6 +1,5 @@
 import { AcquiredDevice } from '@suite-types';
 import { Discovery } from '@wallet-reducers/discoveryReducer';
-import { CONNECT_INIT_ACTION_TYPE } from '@suite-common/connect-init';
 
 /**
  * Strip unserializable fields from Discovery (eg. promises)
@@ -24,13 +23,3 @@ export const serializeDevice = (device: AcquiredDevice, forceRemember?: true) =>
     if (forceRemember) sd.forceRemember = true;
     return sd;
 };
-
-/**
- * Discard @suite-common/connect-init/init actions, we don't care about it in this test
- * if store dispatched these actions.
- * @suite-common/connect-init/init/pending
- * @suite-common/connect-init/init/fulfilled
- * @suite-common/connect-init/init/rejected
- */
-export const discardMockedConnectInitActions = (actions: any[]) =>
-    actions.filter(action => !action.type.includes(CONNECT_INIT_ACTION_TYPE));
