+++ suite-common/connect-init/package.json
@@ -5,7 +5,6 @@
     "license": "See LICENSE.md in repo root",
     "sideEffects": false,
     "main": "src/index",
-    "type": "module",
     "scripts": {
         "lint": "eslint '**/*.{ts,tsx,js}'",
         "type-check": "tsc --build tsconfig.json",
@@ -16,7 +15,7 @@
     "dependencies": {
         "@reduxjs/toolkit": "^1.8.3",
         "@suite-native/state": "*",
-        "@trezor/connect": "9.0.0-beta.4",
+        "@trezor/connect": "9.0.0-beta.5",
         "@trezor/utils": "*"
     },
     "devDependencies": {
