+++ suite-common/connect-init/src/index.ts
@@ -1,1 +1,1 @@
-export * from './actions/prepareConnectInit';
+export * from './slice';
