+++ suite-native/app/src/App.tsx
@@ -1,17 +1,18 @@
-import React, { useEffect } from 'react';
+import React, { useCallback, useEffect } from 'react';
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { Provider, useDispatch, useSelector } from 'react-redux';
 
 import { createAction } from '@reduxjs/toolkit';
 import { NavigationContainer } from '@react-navigation/native';
-import { prepareConnectInit } from '@suite-common/connect-init';
 
+import { prepareConnectInitThunk } from '@suite-common/connect-init';
 import {
     OnboardingStackNavigator,
     selectIsOnboardingFinished,
 } from '@suite-native/module-onboarding';
 import { store } from '@suite-native/state';
+import TrezorConnect from '@trezor/connect';
 
 import { RootTabNavigator } from './navigation/RootTabNavigator';
 import { StylesProvider } from './StylesProvider';
@@ -20,15 +21,30 @@ const AppComponent = () => {
     const dispatch = useDispatch();
     const isOnboardingFinished = useSelector(selectIsOnboardingFinished);
 
+    const getAccountInfo = useCallback(() => {
+        TrezorConnect.getAccountInfo({
+            coin: 'btc',
+            descriptor:
+                'zpub6rszzdAK6RuafeRwyN8z1cgWcXCuKbLmjjfnrW4fWKtcoXQ8787214pNJjnBG5UATyghuNzjn6Lfp5k5xymrLFJnCy46bMYJPyZsbpFGagT',
+        })
+            .then(accountInfo => {
+                // eslint-disable-next-line no-console
+                console.log('Account info result: ', JSON.stringify(accountInfo, null, 2));
+            })
+            .catch(error => {
+                // eslint-disable-next-line no-console
+                console.log('getAccountInfo failed: ', JSON.stringify(error));
+            });
+    }, []);
+
     useEffect(() => {
         const noOperation = createAction('noOperation');
 
+        // TODO handle possible error
         dispatch(
-            prepareConnectInit({
+            prepareConnectInitThunk({
                 actions: {
                     lockDevice: noOperation,
-                    setInitConnectError: noOperation,
-                    onConnectInitialized: noOperation,
                 },
                 selectors: {
                     selectEnabledNetworks: () => [],
@@ -43,7 +59,10 @@ const AppComponent = () => {
                 },
             }),
         );
-    }, [dispatch]);
+
+        // FIXME: remove later - only for testing purposes if we can get account info result.
+        getAccountInfo();
+    }, [dispatch, getAccountInfo]);
 
     if (!isOnboardingFinished) {
         return <OnboardingStackNavigator />;
