+++ suite-native/app/tsconfig.json
@@ -11,6 +11,7 @@
         { "path": "../module-settings" },
         { "path": "../navigation" },
         { "path": "../state" },
+        { "path": "../../packages/connect" },
         { "path": "../../packages/styles" },
         { "path": "../../packages/theme" },
         {
